requirements.txt
flask
flask-sqlalchemy
ollama
